huge refactor with new structure
huge refactor with new structure
