+++ .env
@@ -6,4 +6,4 @@ DB_NAME=sales
 DB_HOST=db
 DB_PORT=5432
 DB_SSLMODE=disable
-MIGRATIONS_PATH=migrations
+MIGRATIONS_PATH=migration
