+++ internal/user/handler/handler.go
@@ -1,12 +1,11 @@
-package route
+package handler
 
 import (
 	"github.com/gofiber/fiber/v2"
-	"github.com/sebajax/go-vertical-slice-architecture/internal/user/api/handler"
 	"github.com/sebajax/go-vertical-slice-architecture/internal/user/service"
 )
 
 // UserRouter is the Router for GoFiber App
 func UserRouter(app fiber.Router, s *service.UserService) {
-	app.Post("/", handler.CreateUser(s.CreateUserServiceProvider))
+	app.Post("/", CreateUser(s.CreateUserServiceProvider))
 }
