+++ pkg/middleware/errorHandler.go
@@ -1,10 +1,11 @@
-package middlewares
+package middleware
 
 import (
 	"github.com/gofiber/fiber/v2"
 	"github.com/gofiber/fiber/v2/log"
+	"github.com/sebajax/go-architecture-angrycoders/pkg/messages"
 	"github.com/sebajax/go-vertical-slice-architecture/pkg/apperror"
-	"github.com/sebajax/go-vertical-slice-architecture/pkg/messages"
+	"github.com/sebajax/go-vertical-slice-architecture/pkg/message"
 )
 
 // ErrorHandler is a middleware that converts AppError to fiber.Error.
@@ -16,13 +17,13 @@ func ErrorHandler(c *fiber.Ctx) error {
 	if err != nil {
 		// Log the error, handle it, or send a custom response
 		if e, ok := err.(*apperror.AppError); ok {
-			log.Error(messages.ErrorResponse(e))
+			log.Error(message.ErrorResponse(e))
 			return c.Status(e.Code).JSON(messages.ErrorResponse(e))
 		}
 
 		// An internal server error ocurred trying to cast error to apperror
-		log.Error(messages.ErrorResponse(err))
-		return c.Status(fiber.StatusInternalServerError).JSON(messages.ErrorResponse(err))
+		log.Error(message.ErrorResponse(err))
+		return c.Status(fiber.StatusInternalServerError).JSON(message.ErrorResponse(err))
 	}
 
 	// If no error, continue the chain
